advanced task

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int fib1 = 0, fib2 = 1, fib3, num = 50, count = 0;
+	long int fib1 = 0, fib2 = 1, fib3, num = 50, count = 0;
 
 	while (count <= num)
 	{
